requirements.txt
python-dotenv==1.1.1
streamlit
boto3
